requirements/base.txt
cx_Oracle==8.3.0
pandas==2.2.2
python-dotenv==1.0.1
